service-tool/requirements.txt
pyserial>=3.5
esptool
